requirement.txt
certifi==2020.6.20
chardet==3.0.4
Django==1.11
djangorestframework==3.11.0
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
